GiveGrant refactored to business logic layer
