+++ api/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "test": "sh test.sh"
+    "test": "node test.js"
   },
   "author": "",
   "license": "ISC",
