+++ api/resolvers/index.js
@@ -694,95 +694,9 @@ const resolvers = {
     giveGrant: async (
       parent,
       { eventId, dreamId, value },
-      { currentUser, models: { Member, Grant, Event, Dream } }
+      { currentUser, controller }
     ) => {
-      const currentMember = await Member.findOne({
-        userId: currentUser.id,
-        eventId,
-      });
-
-      if (!currentMember || !currentMember.isApproved)
-        throw new Error(
-          'You need to be a logged in approved member to grant things'
-        );
-
-      if (value <= 0) throw new Error('Value needs to be more than zero');
-
-      const event = await Event.findOne({ _id: eventId });
-
-      // Check that granting is open
-      if (!event.grantingIsOpen) throw new Error('Granting is not open');
-
-      const dream = await Dream.findOne({ _id: dreamId, eventId });
-
-      if (!dream.approved)
-        throw new Error('Dream is not approved for granting');
-
-      // Check that the max goal of the dream is not exceeded
-      const [
-        { grantsForDream } = { grantsForDream: 0 },
-      ] = await Grant.aggregate([
-        { $match: { dreamId: mongoose.Types.ObjectId(dreamId) } },
-        { $group: { _id: null, grantsForDream: { $sum: '$value' } } },
-      ]);
-
-      // TODO: Create virtual on dream model instead?
-      const maxGoalGrants = Math.ceil(
-        Math.max(dream.maxGoal, dream.minGoal) / event.grantValue
-      );
-
-      if (grantsForDream + value > maxGoalGrants)
-        throw new Error("You can't overfund this dream.");
-
-      // Check that it is not more than is allowed per dream (if this number is set)
-      if (event.maxGrantsToDream && value > event.maxGrantsToDream) {
-        throw new Error(
-          `You can give a maximum of ${event.maxGrantsToDream} grants to one dream`
-        );
-      }
-
-      // Check that user has not spent more grants than he has
-      const [
-        { grantsFromUser } = { grantsFromUser: 0 },
-      ] = await Grant.aggregate([
-        {
-          $match: {
-            memberId: mongoose.Types.ObjectId(currentMember.id),
-            type: 'USER',
-          },
-        },
-        { $group: { _id: null, grantsFromUser: { $sum: '$value' } } },
-      ]);
-
-      if (grantsFromUser + value > event.grantsPerMember)
-        throw new Error('You are trying to spend too many grants.');
-
-      // Check that total budget of event will not be exceeded
-      const [
-        { grantsFromEverybody } = { grantsFromEverybody: 0 },
-      ] = await Grant.aggregate([
-        {
-          $match: {
-            eventId: mongoose.Types.ObjectId(currentMember.eventId),
-            reclaimed: false,
-          },
-        },
-        { $group: { _id: null, grantsFromEverybody: { $sum: '$value' } } },
-      ]);
-
-      const totalGrantsToSpend = Math.floor(
-        event.totalBudget / event.grantValue
-      );
-
-      if (grantsFromEverybody + value > totalGrantsToSpend)
-        throw new Error('Total budget of event is exeeced with this grant');
-
-      return new Grant({
-        eventId: currentMember.eventId,
-        dreamId,
-        value,
-        memberId: currentMember.id,
-      }).save();
+      return await controller.giveGrant(eventId, dreamId, currentUser, value);
     },
     deleteGrant: async (
       parent,
